requirements.txt
streamlit>=1.35.0
pdfplumber==0.11.0
numpy==1.26.4
scipy==1.13.0
pandas==2.2.2
plotly==5.22.0
requests==2.32.3
google-generativeai==0.5.4
python-dotenv==1.0.1
deep-translator==1.11.4
faster-whisper==1.0.3
